top bot bound

+++ IdealPancake/Assets/Scripts/Information.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class Information {
     
     public const int START_X_POS = -10;
-    public const int Y_MAX = 5;
-    public const int Y_MIN = -5;
+    public const float Y_MAX = 5.0f;
+    public const float Y_MIN = -5.0f;
 
     public int bearsEaten = 0;
     public int houseLit = 0;
